feat: 右カメラ用のパラメータ準備
